+++ TheWitnessSolver/PuzzleElement.h
@@ -1,41 +1,10 @@
 #pragma once
 #include "stdafx.h"
 
+#include "Vector2.h"
 #include <iostream>
 #include <unordered_set>
 #include <vector>
-
-// A 2d vector.
-struct Vector2 {
-  Vector2() {}
-  Vector2(int r, int c) : r(r), c(c) {}
-
-  Vector2 operator+(const Vector2& other) {
-    return Vector2(r + other.r, c + other.c);
-  }
-  Vector2 operator-(const Vector2& other) {
-    return Vector2(r - other.r, c - other.c);
-  }
-  void operator=(const Vector2& other) {
-    r = other.r;
-    c = other.c;
-    return;
-  }
-  bool operator==(const Vector2& other) {
-    return (r == other.r) && (c == other.c);
-  }
-  bool operator!=(const Vector2& other) {
-    return !(*this == other);
-  }
-  friend std::ostream& operator<< (std::ostream &out, const Vector2& vec) {
-    out << "[" << vec.r << ", " << vec.c << "]";
-    return out;
-  }
-
-  int r;
-  int c;
-};
-typedef std::vector<Vector2> Vector2List;
 
 // A node on the node matrix.
 // Contains coordinates, its reachable naighbors, and some extra info.
@@ -62,8 +31,8 @@ struct Node {
   bool isTail;
   bool onEdge;
 };
-typedef std::unordered_set<Node*> NodeSet;
-typedef std::vector<Node*> NodeVector;
+typedef std::unordered_set<Node*> NodePtrSet;
+typedef std::vector<Node*>        NodePtrVector;
 typedef std::vector<std::vector<Node>> NodeMatrix;
 
 // Block type
@@ -83,7 +52,7 @@ struct Block {
   }
   void InitBlock(int r, int c) {
     coord = Vector2(r, c);
-    type  = Empty;
+    type = Empty;
     clusterID = -1;
     neighborSet.clear();
     visited = false;
@@ -95,9 +64,9 @@ struct Block {
   int clusterID;
   bool visited; // used by segmentation
 };
+typedef std::unordered_set<Block*>      BlockPtrSet;
+typedef std::vector<BlockPtrSet>        BlockPtrSetVector;
 typedef std::vector<std::vector<Block>> BlockMatrix;
-typedef std::unordered_set<Block*> BlockSet;
-typedef std::vector<BlockSet> BlockSetVector;
 
 // A side is a line connecting 2 adjacent nodes
 // When initialized, the order of node1 & node2 is rearranged
@@ -129,7 +98,7 @@ struct Side {
   // TODO: we only need to check the first case now
   bool operator==(const Side& other) const {
     if ((node1->coord == other.node1->coord && node2->coord == other.node2->coord) ||
-        (node1->coord == other.node2->coord && node2->coord == other.node1->coord)) {
+      (node1->coord == other.node2->coord && node2->coord == other.node1->coord)) {
       return true;
     }
     else return false;
